Comissão > vincular servidor (tela reformulada)

+++ app/Models/Regras/ServidorComissaoRegras.php
@@ -9,18 +9,22 @@ class ServidorComissaoRegras
 {
     public static function salvar($dados)
     {
+
         //caso o servidor ja tenha alguma comissão vinculada, este registro é deletado e é efetuado um novo registro
         $servidorNaComissao = ServidorComissao::where('fk_servidor', '=', $dados->fk_servidor)
                                                 ->whereNull('deleted_at')
                                                 ->first();
         if($servidorNaComissao) {
-            $servidorNaComissao->delete();
+            if(!isset($dados->fk_comissao)){
+                $servidorNaComissao->delete();
+            }
+        } else {
+            $servidorComissao = new ServidorComissao;
+            $servidorComissao->fk_servidor = $dados->fk_servidor;
+            $servidorComissao->fk_comissao = $dados->fk_comissao;
+            $servidorComissao->save();
         }
 
-        $servidorComissao = new ServidorComissao;
-        $servidorComissao->fk_servidor = $dados->fk_servidor;
-        $servidorComissao->fk_comissao = $dados->fk_comissao;
-        $servidorComissao->save();
     }
 
 }
